# enginepb: clean up MVCCValueHeaderPure hack