--- a/pkg/kv/kvnemesis/kvnemesisutil/seq.go
+++ b/pkg/kv/kvnemesis/kvnemesisutil/seq.go
@@ -27,6 +27,12 @@ type SeqContainer uint32
 
 var _ protoutil.Message = (*SeqContainer)(nil)
 
+// IsEmpty is used by generated marshalling code (with gogoproto.omitempty
+// extension).
+func (m SeqContainer) IsEmpty() bool {
+	return m == 0
+}
+
 // Reset implements (a part of) protoutil.Message.
 func (m *SeqContainer) Reset() {
 	*m = 0
@@ -108,6 +114,8 @@ func (m SeqContainer) Get() Seq {
 // proto messages. It uses no space. When the crdb_test build tag is set, this
 // type is instead represented by a NoopContainer.
 type NoopContainer struct{}
+
+func (m *NoopContainer) IsEmpty() bool { return true }
 
 func (m *NoopContainer) Reset() {}