--- a/pkg/util/hlc/timestamp.go
+++ b/pkg/util/hlc/timestamp.go
@@ -187,6 +187,8 @@ func (t Timestamp) AsOfSystemTime() string {
 // IsEmpty returns true if t is an empty Timestamp.
 // gcassert:inline
 func (t Timestamp) IsEmpty() bool {
+	// TODO(radu): consider making timestamps with zero wall time and non-zero
+	// logical time illegal. Then we can check just the wall time.
 	return t == Timestamp{}
 }